allow toggling between light and dark themes
uses javascript and cookies to be persistent. Requires a load cookie command in body text, might alter load speed slightly?
Alternate css using style switching in the browser is available too (View->Styles in Firefox) but this is not supported by Safari or Chrome, nor is it persistent.  Not sure about placement of toggle buttons, currently in bottom right above footer


Former-commit-id: 8f27a0c411315667be65ebc9e80391a1f8b36b3f