addresses the header html problem by hardcoding a template
not optimal, preferably a template file would only be used when the option was given in the _config.yaml
in this solution, template file is always used.  But net effect accomplishes goal of providing data URIs when
they are toggled on in _config.yaml and not providing them when they are toggled of in _config.yaml. Closes #58


Former-commit-id: 8f664cc859f4977840076026e5db1019f75cb562