Successfully implemented local site search, closes #7
Uses Marran's method.  Search is by term only, word
proximity or phrases cannot be searched for.  Uses
javascript, ruby, liquid, and css to create solution.

Search is rather fast on generation and very fast once deployed
on the site, thanks to the stemming tricks involved.

Documentation of the approach still needs to be added to the
plugin file.  Best to switch search generation off when
generating an archival version.


Former-commit-id: 94abd08b553ef40bff8eecedfabf8d6ce5eb492e